Fix the test case broken.
