+++ osu.Game.Rulesets.Karaoke.Tests/Editor/Lyrics/CaretPosition/Algorithms/BaseCaretPositionAlgorithmTest.cs
@@ -33,8 +33,8 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Lyrics.CaretPosition.Algorithms
             invokeAlgorithm?.Invoke(algorithm);
 
             var actual = algorithm.MoveUp(caret);
-            assertEqual(expected, actual);
-            checkCaretGenerateType(CaretGenerateType.Action, actual);
+            AssertEqual(expected, actual);
+            CheckCaretGenerateType(CaretGenerateType.Action, actual);
         }
 
         protected void TestMoveDown(Lyric[] lyrics, TCaret caret, TCaret? expected, Action<TAlgorithm>? invokeAlgorithm = null)
@@ -46,34 +46,8 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Lyrics.CaretPosition.Algorithms
             invokeAlgorithm?.Invoke(algorithm);
 
             var actual = algorithm.MoveDown(caret);
-            assertEqual(expected, actual);
-            checkCaretGenerateType(CaretGenerateType.Action, actual);
-        }
-
-        protected void TestMoveLeft(Lyric[] lyrics, TCaret caret, TCaret? expected, Action<TAlgorithm>? invokeAlgorithm = null)
-        {
-            var algorithm = (TAlgorithm?)Activator.CreateInstance(typeof(TAlgorithm), new object[] { lyrics });
-            if (algorithm == null)
-                throw new ArgumentNullException();
-
-            invokeAlgorithm?.Invoke(algorithm);
-
-            var actual = algorithm.MoveLeft(caret);
-            assertEqual(expected, actual);
-            checkCaretGenerateType(CaretGenerateType.Action, actual);
-        }
-
-        protected void TestMoveRight(Lyric[] lyrics, TCaret caret, TCaret? expected, Action<TAlgorithm>? invokeAlgorithm = null)
-        {
-            var algorithm = (TAlgorithm?)Activator.CreateInstance(typeof(TAlgorithm), new object[] { lyrics });
-            if (algorithm == null)
-                throw new ArgumentNullException();
-
-            invokeAlgorithm?.Invoke(algorithm);
-
-            var actual = algorithm.MoveRight(caret);
-            assertEqual(expected, actual);
-            checkCaretGenerateType(CaretGenerateType.Action, actual);
+            AssertEqual(expected, actual);
+            CheckCaretGenerateType(CaretGenerateType.Action, actual);
         }
 
         protected void TestMoveToFirst(Lyric[] lyrics, TCaret? expected, Action<TAlgorithm>? invokeAlgorithm = null)
@@ -85,8 +59,8 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Lyrics.CaretPosition.Algorithms
             invokeAlgorithm?.Invoke(algorithm);
 
             var actual = algorithm.MoveToFirst();
-            assertEqual(expected, actual);
-            checkCaretGenerateType(CaretGenerateType.Action, actual);
+            AssertEqual(expected, actual);
+            CheckCaretGenerateType(CaretGenerateType.Action, actual);
         }
 
         protected void TestMoveToLast(Lyric[] lyrics, TCaret? expected, Action<TAlgorithm>? invokeAlgorithm = null)
@@ -98,8 +72,8 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Lyrics.CaretPosition.Algorithms
             invokeAlgorithm?.Invoke(algorithm);
 
             var actual = algorithm.MoveToLast();
-            assertEqual(expected, actual);
-            checkCaretGenerateType(CaretGenerateType.Action, actual);
+            AssertEqual(expected, actual);
+            CheckCaretGenerateType(CaretGenerateType.Action, actual);
         }
 
         protected void TestMoveToTarget(Lyric[] lyrics, Lyric lyric, TCaret? expected, Action<TAlgorithm>? invokeAlgorithm = null)
@@ -111,11 +85,11 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Lyrics.CaretPosition.Algorithms
             invokeAlgorithm?.Invoke(algorithm);
 
             var actual = algorithm.MoveToTarget(lyric);
-            assertEqual(expected, actual);
-            checkCaretGenerateType(CaretGenerateType.TargetLyric, actual);
+            AssertEqual(expected, actual);
+            CheckCaretGenerateType(CaretGenerateType.TargetLyric, actual);
         }
 
-        private void assertEqual(TCaret? expected, TCaret? actual)
+        protected void AssertEqual(TCaret? expected, TCaret? actual)
         {
             if (expected == null || actual == null)
             {
@@ -128,7 +102,7 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Lyrics.CaretPosition.Algorithms
             }
         }
 
-        private void checkCaretGenerateType(CaretGenerateType expected, TCaret? actual)
+        protected void CheckCaretGenerateType(CaretGenerateType expected, TCaret? actual)
         {
             if (actual == null)
                 return;
